pyproject.toml
[tool.poetry]
name = "sushi"
version = "0.1.0"
description = ""
authors = ["Thenuja Viknarajah <thenuja_1@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
azure-identity = "^1.12.0"
azure-keyvault-secrets = "^4.6.0"
azure-storage-blob = "^12.14.1"
paramiko = "^3.0.0"
pydantic = "^1.10.4"
pendulum = "^2.1.2"

[tool.poetry.scripts]
sushictl = "sushi.cli:cli"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
